Add a hurry flag to the workflow spec.
The Slurm WLM has the concept of entering teardown state with the help of
a "hurry" flag. This flag indicates that any in-progress data transfers must
be terminated, or that any transfers not yet begun must be skipped.

It is left to the DWS driver implementation to determine how, and whether, to
implement this flag.

Add constraints to the webhook to allow the hurry flag to be used only with
teardown state.

Hook up the webhook in the controller tests. We need the reconciler running
with the webhook to test things in teardown state.

Hook up tests in dockerfile so they run during the github workflow action, and
enable tests in the workflow.

+++ api/v1alpha1/workflow_webhook_test.go
@@ -21,10 +21,13 @@ package v1alpha1
 
 import (
 	"context"
+	"fmt"
 
+	"github.com/google/uuid"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // These tests are written in BDD-style using Ginkgo framework. Refer to
@@ -36,9 +39,10 @@ var _ = Describe("Workflow Webhook", func() {
 	)
 
 	BeforeEach(func() {
+		wfid := uuid.NewString()[0:8]
 		workflow = &Workflow{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:      "test",
+				Name:      fmt.Sprintf("w%s", wfid),
 				Namespace: metav1.NamespaceDefault,
 			},
 			Spec: WorkflowSpec{
@@ -46,17 +50,32 @@ var _ = Describe("Workflow Webhook", func() {
 				DWDirectives: []string{},
 			},
 		}
-
-		Expect(k8sClient.Create(context.TODO(), workflow)).To(Succeed())
 	})
 
 	AfterEach(func() {
-		Expect(k8sClient.Delete(context.TODO(), workflow)).To(Succeed())
+		if workflow != nil {
+			Expect(k8sClient.Delete(context.TODO(), workflow)).To(Succeed())
+		}
 	})
 
 	It("should have workflow environmental variables set successfully", func() {
+		Expect(k8sClient.Create(context.TODO(), workflow)).To(Succeed())
 		Expect(workflow.Status.Env).To(HaveKeyWithValue("DW_WORKFLOW_NAME", workflow.Name))
 		Expect(workflow.Status.Env).To(HaveKeyWithValue("DW_WORKFLOW_NAMESPACE", workflow.Namespace))
 	})
 
+	It("Fails to create workflow with hurry flag set", func() {
+		workflow.Spec.Hurry = true
+		Expect(k8sClient.Create(context.TODO(), workflow)).ToNot(Succeed())
+		workflow = nil
+	})
+
+	It("Create workflow, then fail to update hurry flag", func() {
+		Expect(k8sClient.Create(context.TODO(), workflow)).To(Succeed())
+		Eventually(func(g Gomega) {
+			g.Expect(k8sClient.Get(context.TODO(), client.ObjectKeyFromObject(workflow), workflow)).To(Succeed())
+		}).Should(Succeed(), "get new workflow")
+		workflow.Spec.Hurry = true
+		Expect(k8sClient.Update(context.TODO(), workflow)).ToNot(Succeed())
+	})
 })
